suggestion_rework
