+++ react-training/src/day4/Ques12/mock.js
@@ -1,0 +1,58 @@
+export const createData = (id, name, calories, fat, carbs, protein) => ({
+    id,
+    name,
+    calories,
+    fat,
+    carbs,
+    protein,
+  });
+  
+  export const rows = [
+    createData(1, "Cupcake", 305, 3.7, 67, 4.3),
+    createData(2, "Donut", 452, 25.0, 51, 4.9),
+    createData(3, "Eclair", 262, 16.0, 24, 6.0),
+    createData(4, "Frozen yoghurt", 159, 6.0, 24, 4.0),
+    createData(5, "Gingerbread", 356, 16.0, 49, 3.9),
+    createData(6, "Honeycomb", 408, 3.2, 87, 6.5),
+    createData(7, "Ice cream sandwich", 237, 9.0, 37, 4.3),
+    createData(8, "Jelly Bean", 375, 0.0, 94, 0.0),
+    createData(9, "KitKat", 518, 26.0, 65, 7.0),
+    createData(10, "Lollipop", 392, 0.2, 98, 0.0),
+    createData(11, "Marshmallow", 318, 0, 81, 2.0),
+    createData(12, "Nougat", 360, 19.0, 9, 37.0),
+    createData(13, "Oreo", 437, 18.0, 63, 4.0),
+  ];
+  
+  export const headCells = [
+    {
+      id: "name",
+      numeric: false,
+      disablePadding: true,
+      label: "Dessert (100g serving)",
+    },
+    {
+      id: "calories",
+      numeric: true,
+      disablePadding: false,
+      label: "Calories",
+    },
+    {
+      id: "fat",
+      numeric: true,
+      disablePadding: false,
+      label: "Fat (g)",
+    },
+    {
+      id: "carbs",
+      numeric: true,
+      disablePadding: false,
+      label: "Carbs (g)",
+    },
+    {
+      id: "protein",
+      numeric: true,
+      disablePadding: false,
+      label: "Protein (g)",
+    },
+  ];
+
